% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_models.R
\name{add_models}
\alias{add_models}
\title{Add nuisance models to a study task}
\usage{
add_models(task, ...)
}
\arguments{
\item{task}{A \code{enact_task} object.}

\item{\dots}{Selector objects created by \code{\link{treatments}},
\code{\link{outcomes}}, \code{\link{censoring}}, and/or \code{\link{mtp}}.}
}
\value{
The modified \code{enact_task}.
}
\description{
Specifies the learners and metalearners for treatment mechanisms, outcome
regressions, censoring mechanisms, and/or MTP density ratio models.  Must be
called after \code{\link{add}} and \code{\link{add_cv_folds}}.  Can only be
called once per task.
}
