% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_inference.R
\name{all_outcomes}
\alias{all_outcomes}
\title{Select all outcomes for do_tmle}
\usage{
all_outcomes()
}
\value{
\code{NULL}
}
\description{
Convenience function for the \code{which} argument of \code{\link{do_tmle}}.
Returns \code{NULL}, which is interpreted as "all outcomes".
}
