% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_models.R
\name{censoring}
\alias{censoring}
\title{Specify censoring mechanism models}
\usage{
censoring(learners, metalearners = NULL)
}
\arguments{
\item{learners}{A single \code{enfold_learner} (or pipeline/grid/list), or a
list of such objects.  Required.}

\item{metalearners}{A single \code{enfold_learner} created by a metalearner
constructor (e.g. \code{mtl_superlearner()}).  Required when
\code{learners} is a list with multiple elements.}
}
\value{
An object of class \code{enact_selector_censoring}.
}
\description{
Selector for \code{\link{add_models}}.  Targets censoring mechanism models
for all outcomes that have censoring indicators.  The same learner
specification is used for each censoring mechanism.
}
