% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enact_demo}
\alias{enact_demo}
\title{Example dataset for enact}
\format{
A data frame with TODO rows and TODO variables:
\describe{
\item{TODO}{TODO description}
}
}
\source{
Simulated data.
}
\usage{
enact_demo
}
\description{
A simulated dataset illustrating a typical causal inference study setup
compatible with the enact workflow.
}
\keyword{datasets}
