% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{log_relative_ate}
\alias{log_relative_ate}
\title{Log relative average treatment effect contrast}
\usage{
log_relative_ate()
}
\value{
A function with a \code{"label"} attribute set to \code{"Log Relative ATE"}.
}
\description{
Returns a function \code{function(reference, treatment)} that computes
\code{log(treatment) - log(reference)}.  Pass to the \code{contrasts}
argument of \code{\link{add_outcome}}.
}
