% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_models.R
\name{mtp}
\alias{mtp}
\title{Specify MTP density ratio models}
\usage{
mtp(learners, metalearner = NULL)
}
\arguments{
\item{learners}{A single \code{enfold_learner} (or pipeline/grid/list), or a
list of such objects.  Required.}

\item{metalearner}{A single \code{enfold_learner} created by a metalearner
constructor (e.g. \code{mtl_superlearner()}).  Required when
\code{learners} is a list with multiple elements.}
}
\value{
An object of class \code{enact_selector_mtp}.
}
\description{
Selector for \code{\link{add_models}}.  Targets the 2N density ratio
classifier for Modified Treatment Policies.
}
