% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.R
\name{strip}
\alias{strip}
\title{Strip everything but the reporting-relevant slots from an enact_task}
\usage{
strip(task, ...)
}
\arguments{
\item{task}{A study task object (e.g. \code{enact_task}).}

\item{\dots}{Additional arguments.}
}
\value{
The stripped task.
}
\description{
Generic function for removing slots from an \code{enact_task}
that are not needed for reporting. This file is safe to send to collaborators
without privacy concerns, or to reduce memory footprint before reporting.
}
